Use a lepton collection in the AK8 Wjj selector class
In inclusive analysis the AK8LS jets are selected such that the lepton used in the AK8LS cuts is the closest lepton to a given AK8LS jet. This implies that the selector class sees the full loose lepton collection and not a single lepton lepton defined by event selection cuts. In inclusive analysis we cannot really impose any such requirements on the leptons.