Remove incomplete samples
There are three samples that contain too few events compared to the numbers on DBS:
/GluGluToRadionToHHTo2V2Tau_M-850_narrow_13TeV-madgraph_correctedcfg/RunIIFall17MiniAODv2-PU2017_12Apr2018_94X_mc2017_realistic_v14-v1/MINIAODSIM 5.7%
/GluGluToBulkGravitonToHHTo2V2Tau_M-270_narrow_13TeV-madgraph_correctedcfg/RunIIFall17MiniAODv2-PU2017_12Apr2018_94X_mc2017_realistic_v14-v1/MINIAODSIM 4.5%
/GluGluToHHTo2V2Tau_node_11_13TeV-madgraph_correctedcfg/RunIIFall17MiniAODv2-PU2017_12Apr2018_94X_mc2017_realistic_v14-v1/MINIAODSIM 0.8%
The CRAB tasks have been successfully finished for these datasets, however, if the dataset is still in "production" at the time of the submission, only events available at the time of submission are considered in Ntupelization. Apaprently, these samples suffered this very problem and we have to reprocess them again now that they all have been marked as "valid" samples.