Disable VBF samples
As it was discussed offline, the VBF category will be skipped for this iteration of the analysis because its low sensitivity, and also because there's no reweighting scheme developed for the non-resonant VBF samples, yet. For these reasons, it's best to exclude the VBF sample in the analysis in order to reduce unnecessary load on our computing infrastructure.